+++ R/get_marginalcontrasts.R
@@ -100,12 +100,12 @@ get_marginalcontrasts <- function(model,
       verbose = verbose,
       ...
     )
-  } else if (identical(comparison, "inequality")) {
+  } else if (identical(comparison, "inequality") || identical(comparison, "total")) {
     # inequality effect summary, see Trenton D. Mize, Bing Han 2025
     # Inequality and Total Effect Summary Measures for Nominal and Ordinal Variables
     # Sociological Science February 5, 10.15195/v12.a7
     # this requires a special handling, because we can only use it with avg_comparisons
-    out <- .calculate_inequality_effect(model, model_data, my_args, ci, ...)
+    out <- .calculate_inequality_effect(model, model_data, my_args, comparison, ci, ...)
   } else {
     # for contrasts of categorical predictors, we call avg_predictions
     out <- estimate_means(
@@ -165,7 +165,7 @@ get_marginalcontrasts <- function(model,
 
 # special contrasts: inequality---------------- -------------------------------
 
-.calculate_inequality_effect <- function(model, model_data, my_args, ci, ...) {
+.calculate_inequality_effect <- function(model, model_data, my_args, comparison, ci, ...) {
   # to calcualte marginal effects inequalities, all contrast predictors
   # must be factors
   check_factors <- .safe(vapply(model_data[my_args$contrast], is.factor, logical(1)), NULL)
@@ -184,6 +184,15 @@ get_marginalcontrasts <- function(model,
     hypothesis = ~I(mean(abs(x))) | term,
     ...
   )
+  # for the total marginal effects, we need to call "hypothesis()" again, this
+  # time with ~revpairwise option
+  if (comparison == "total") {
+    if (nrow(out) < 2) {
+      insight::format_error("Total marginal effects can only be calculated for more than one marginal effect inequalty.")
+    }
+    out <- marginaleffects::hypotheses(out, hypothesis = ~revpairwise)
+  }
+
   class(out) <- unique(c("marginaleffects_means", class(out)))
   format(out, model, ci, hypothesis = "inequality", ...)
 }
@@ -394,7 +403,7 @@ get_marginalcontrasts <- function(model,
   c(
     "pairwise", "reference", "sequential", "meandev", "meanotherdev",
     "revpairwise", "revreference", "revsequential", "poly", "helmert",
-    "trt_vs_ctrl", "joint", "inequality"
+    "trt_vs_ctrl", "joint", "inequality", "total"
   )
 }