+++ NEWS.md
@@ -2,9 +2,11 @@
 
 ## Changes
 
-* The `comparison` argument gets a new option, `"inequality"`, to compute the
-  marginal effects inequality measure, which summarizes the holistic effect of
-  categorical predictors.
+* The `comparison` argument gets two new option, `"inequality"` and `"total"`,
+  to compute the marginal effects inequality measure, which summarizes the the
+  overall effect of categorical predictors or the comprehensive effect of a
+  predictor across all outcome categories of a nominal or ordinal dependent
+  variable.
 
 * Added docs to show how to use *modelbased* with finite mixture models from
   package *brms*.